# Server/swagger-documentation.yaml
openapi: 3.0.3
info:
  title: CashHaven API
  description: |
    API completa para o sistema CashHaven - Sistema de gestão comercial com funcionalidades de vendas, estoque, clientes, pedidos e integração com WhatsApp.
    
    ## Autenticação
    A API utiliza autenticação JWT (JSON Web Token). Para acessar endpoints protegidos, inclua o token no header Authorization:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## Funcionalidades Principais
    - **Autenticação e Autorização**: Login/registro de usuários com diferentes níveis de acesso
    - **Gestão de Usuários**: CRUD completo para usuários do sistema
    - **Gestão de Clientes**: Cadastro e gerenciamento de clientes
    - **Controle de Estoque**: Gerenciamento completo de produtos e estoque
    - **Sistema de Vendas**: Processamento de vendas e transações
    - **Gestão de Pedidos**: Criação e acompanhamento de pedidos
    - **Integração WhatsApp**: Comunicação automatizada via WhatsApp
    - **Relatórios**: Geração de relatórios de vendas e estoque
    - **WebSocket**: Comunicação em tempo real
    
  version: 1.0.0
  contact:
    name: CashHaven Support
    email: support@cashhaven.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5238/api
    description: Servidor de desenvolvimento
  - url: https://api.cashhaven.com/api
    description: Servidor de produção
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login
  schemas:
...
    LoginRequest:
      type: object
      properties:
        usuario:
          type: string
          description: Nome de usuário
          example: "admin"
        senha:
          type: string
          description: Senha do usuário
          example: "senha123"
      required:
        - usuario
        - senha
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indica se o login foi bem-sucedido
          example: true
        token:
          type: string
          description: Token JWT para autenticação
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiration:
          type: string
          format: date-time
          description: Data de expiração do token
          example: "2024-01-16T10:30:00Z"
        nome:
          type: string
          description: Nome do usuário
          example: "Administrador"
        id:
          type: integer
          description: ID do usuário
          example: 1
        cargo:
          type: string
          description: Cargo do usuário
          example: "Administrador"
    RegisterRequest:
      type: object
      properties:
        nome:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
        nome_usuario:
          type: string
          description: Nome de usuário único
          example: "joao.silva"
        senha:
          type: string
          description: Senha do usuário
          example: "senha123"
        cargo:
          type: string
          description: Cargo/função do usuário
          example: "Vendedor"
        adm:
          type: integer
          description: Nível de administrador (0 ou 1)
          example: 0
      required:
        - nome
        - nome_usuario
        - senha
        - adm
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indica se a operação foi bem-sucedida
          example: true
        message:
          type: string
          description: Mensagem descritiva do resultado
          example: "Operação realizada com sucesso"
        data:
          type: object
          description: Dados retornados pela operação
        error:
          type: string
          description: Mensagem de erro (quando success = false)
        details:
          type: object
          description: Detalhes adicionais sobre o erro
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Mensagem de erro
          example: "Erro interno do servidor"
        error:
          type: string
          description: Tipo do erro
        details:
          type: object
          description: Detalhes técnicos do erro
paths:
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Realizar login no sistema
      description: |
        Autentica um usuário no sistema usando nome de usuário e senha.
        Retorna um token JWT que deve ser usado para acessar endpoints protegidos.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Login como administrador
                value:
                  usuario: "admin"
                  senha: "admin123"
              vendedor:
                summary: Login como vendedor
                value:
                  usuario: "vendedor01"
                  senha: "vend123"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  summary: Login bem-sucedido
                  value:
                    success: true
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    expiration: "2024-01-16T10:30:00Z"
                    nome: "ADMINISTRADOR"
                    id: 1
                    cargo: "Administrador"
        '500':
          description: Erro no login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_credentials:
                  summary: Credenciais inválidas
                  value:
                    success: false
                    errors: ["Usuário não encontrado"]
                wrong_password:
                  summary: Senha incorreta
                  value:
                    success: false
                    errors: ["Senha incorreta"]
  /auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar novo usuário
      description: |
        Cria um novo usuário no sistema. Requer privilégios administrativos.
        A senha é automaticamente criptografada antes de ser armazenada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              new_user:
                summary: Novo usuário vendedor
                value:
                  nome: "Maria Santos"
                  nome_usuario: "maria.santos"
                  senha: "senha123"
                  cargo: "Vendedora"
                  adm: 0
              new_admin:
                summary: Novo administrador
                value:
                  nome: "Carlos Admin"
                  nome_usuario: "carlos.admin"
                  senha: "admin456"
                  cargo: "Gerente"
                  adm: 1
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "Usuário cadastrado com sucesso"
        '500':
          description: Erro ao criar usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...